finding outliers
As discussed, I am attaching the data for the Chartering Project where we need your help. 

 

The Y-variable or the prime variable of our interest has been highlighted in the attachment while all other remaining variables (except for the Date col.) are to be treated as the X-vars. Now, here are the objectives that we want you to engage on – 

·         We have a hypothesis that there might have been disruptions in our Y-variable across the past 11 years of the data as present in the attachment. For example, the Covid period (approx. Dec-19 to May-20) is an example of such a disruption. Similarly, can we determine if there are other windows of such disruptions and if it indeed exists, then how are we mathematically identifying them?

 

·         Can we highlight a set of X-variables which display similar such patterns during the above mentioned disruption period or may be prior or post to that disruption period?

 

Kindly take a look at the data and lets connect if you need any help in understanding the definition of the variables or if you want any help in strategizing your approach.